created baseCard slot for card components to use
