+++ src/components/employee-beta/LanguagesCard.vue
@@ -1,13 +1,6 @@
 <template>
   <div class="infoTab">
-    <v-card>
-      <v-card-title class="d-flex align-center justify-space-between header_style">
-        <h3 class="text-white px-2">Foreign Languages</h3>
-        <v-btn v-if="isAdmin || isUser" density="comfortable" variant="text" icon="">
-          <v-tooltip activator="parent" location="top"> Edit Profile </v-tooltip>
-          <v-icon id="edit" color="white"> mdi-pencil </v-icon>
-        </v-btn>
-      </v-card-title>
+    <base-card title="Languages">
       <v-card-text class="px-7 pt-5 pb-1 text-black">
         <!-- Employee has entered languages-->
         <div v-if="!isEmpty(model.languages)">
@@ -31,15 +24,16 @@
           </v-card-actions>
         </div>
       </v-card-text>
-    </v-card>
-    <languages-modal v-model="toggleModal" :model="model"></languages-modal>
+      <languages-modal v-model="toggleModal" :model="model"></languages-modal>
+    </base-card>
   </div>
 </template>
 
 <script setup>
 import { isEmpty, sortLanguagesByProficiency } from '@/utils/utils';
-import { computed, inject, ref } from 'vue';
+import { computed, ref } from 'vue';
 import LanguagesModal from './modals/LanguagesModal.vue';
+import BaseCard from '@/components/employee-beta/BaseCard.vue';
 
 // |--------------------------------------------------|
 // |                                                  |
@@ -48,9 +42,6 @@ import LanguagesModal from './modals/LanguagesModal.vue';
 // |--------------------------------------------------|
 
 const props = defineProps(['model']);
-
-const isAdmin = inject('isAdmin');
-const isUser = inject('isUser');
 const toggleModal = ref(false);
 
 // |--------------------------------------------------|
